requirements.txt
numpy
tensorboard
tqdm
ConfigArgParse
matplotlib
plotly
shapely
scipy
setuptools
scikit-learn
pillow
pyyaml
pandas
joblib
sentence_transformers
wandb
trimesh
mesh_to_sdf
seaborn
tabulate
einops
torchdiffeq
torch-summary
pygame
gitpython
scikit-video
# functorch
opencv-python
moviepy
hydra-core
urdfpy
pybullet
dirsync
colour
h5py
pre-commit
black
